added `SmTransformer` pipeline that allows custom user state machine transformation!
